+++ src/telemetry.hpp
@@ -10,7 +10,7 @@
 #include <networktables/StructArrayTopic.h>
 #include <networktables/StructTopic.h>
 
-#include "drivetrain.hpp"
+#include "subsystems/drivetrain.hpp"
 
 class Telemetry {
 private:
